// sistem-pelaporan-komunitas/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  MASYARAKAT: 'MASYARAKAT',
  PETUGAS: 'PETUGAS',
  ADMIN: 'ADMIN'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const ReportStatus = {
  MENUNGGU_VERIFIKASI: 'MENUNGGU_VERIFIKASI',
  VERIFIKASI_DITOLAK: 'VERIFIKASI_DITOLAK',
  DIPROSES: 'DIPROSES',
  DITINDAKLANJUTI: 'DITINDAKLANJUTI',
  SELESAI: 'SELESAI'
} as const

export type ReportStatus = (typeof ReportStatus)[keyof typeof ReportStatus]


export const ReportPriority = {
  RENDAH: 'RENDAH',
  SEDANG: 'SEDANG',
  TINGGI: 'TINGGI',
  URGENT: 'URGENT'
} as const

export type ReportPriority = (typeof ReportPriority)[keyof typeof ReportPriority]


export const RegionLevel = {
  KECAMATAN: 'KECAMATAN',
  KELURAHAN: 'KELURAHAN',
  DESA: 'DESA'
} as const

export type RegionLevel = (typeof RegionLevel)[keyof typeof RegionLevel]